+++ myems-api/excelexporters/equipmentefficiency.py
@@ -490,14 +490,14 @@ def generate_excel(report,
                 line.add_data(line_data, titles_from_data=True)
                 line.set_categories(labels)
                 line_data = line.series[0]
-                line_data.marker.symbol = "circle"
+                line_data.marker.symbol = "auto"
                 line_data.smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(chart_start_row_number)
@@ -513,14 +513,14 @@ def generate_excel(report,
                 line.add_data(line_data, titles_from_data=True)
                 line.set_categories(labels)
                 line_data = line.series[0]
-                line_data.marker.symbol = "circle"
+                line_data.marker.symbol = "auto"
                 line_data.smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(chart_start_row_number)
@@ -536,14 +536,14 @@ def generate_excel(report,
                 line.add_data(line_data, titles_from_data=True)
                 line.set_categories(labels)
                 line_data = line.series[0]
-                line_data.marker.symbol = "circle"
+                line_data.marker.symbol = "auto"
                 line_data.smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(chart_start_row_number)
@@ -838,14 +838,14 @@ def generate_excel(report,
                 line.add_data(reporting_line_data, titles_from_data=True)
                 line.set_categories(labels)
                 for j in range(len(line.series)):
-                    line.series[j].marker.symbol = "circle"
+                    line.series[j].marker.symbol = "auto"
                     line.series[j].smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(chart_start_row_number)
@@ -870,14 +870,14 @@ def generate_excel(report,
                 line.add_data(reporting_line_data, titles_from_data=True)
                 line.set_categories(labels)
                 for j in range(len(line.series)):
-                    line.series[j].marker.symbol = "circle"
+                    line.series[j].marker.symbol = "auto"
                     line.series[j].smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(chart_start_row_number)
@@ -902,14 +902,14 @@ def generate_excel(report,
                 line.add_data(reporting_line_data, titles_from_data=True)
                 line.set_categories(labels)
                 for j in range(len(line.series)):
-                    line.series[j].marker.symbol = "circle"
+                    line.series[j].marker.symbol = "auto"
                     line.series[j].smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(chart_start_row_number)
@@ -1064,7 +1064,7 @@ def generate_excel(report,
             line.add_data(line_data, titles_from_data=True)
             line.set_categories(labels)
             line_data = line.series[0]
-            line_data.marker.symbol = "circle"
+            line_data.marker.symbol = "auto"
             line_data.smooth = True
             line.x_axis.crosses = 'min'
             line.height = 8.25