+++ myems-api/excelexporters/shopfloorplan.py
@@ -392,7 +392,7 @@ def generate_excel(report,
         s1 = pie.series[0]
         s1.dLbls = DataLabelList()
         s1.dLbls.showCatName = False
-        s1.dLbls.showVal = True
+        s1.dLbls.showVal = False
         s1.dLbls.showPercent = True
         ws.add_chart(pie, 'E' + str(chart_start_row_number))
 
@@ -461,7 +461,7 @@ def generate_excel(report,
         s1 = pie.series[0]
         s1.dLbls = DataLabelList()
         s1.dLbls.showCatName = False
-        s1.dLbls.showVal = True
+        s1.dLbls.showVal = False
         s1.dLbls.showPercent = True
         ws.add_chart(pie, 'E' + str(chart_start_row_number))
 
@@ -567,14 +567,14 @@ def generate_excel(report,
                     line.add_data(line_data, titles_from_data=True)
                     line.set_categories(labels)
                     line_data = line.series[0]
-                    line_data.marker.symbol = "circle"
+                    line_data.marker.symbol = "auto"
                     line_data.smooth = True
                     line.x_axis.crosses = 'min'
                     line.height = 8.25
                     line.width = 24
                     line.dLbls = DataLabelList()
                     line.dLbls.dLblPos = 't'
-                    line.dLbls.showVal = True
+                    line.dLbls.showVal = False
                     line.dLbls.showPercent = False
                     chart_col = 'B'
                     chart_cell = chart_col + str(table_start_draw_flag + 6 * i)
@@ -740,14 +740,14 @@ def generate_excel(report,
                     line.add_data(reporting_line_data, titles_from_data=True)
                     line.set_categories(labels)
                     for j in range(len(line.series)):
-                        line.series[j].marker.symbol = "circle"
+                        line.series[j].marker.symbol = "auto"
                         line.series[j].smooth = True
                     line.x_axis.crosses = 'min'
                     line.height = 8.25
                     line.width = 24
                     line.dLbls = DataLabelList()
                     line.dLbls.dLblPos = 't'
-                    line.dLbls.showVal = True
+                    line.dLbls.showVal = False
                     line.dLbls.showPercent = False
                     chart_col = 'B'
                     chart_cell = chart_col + str(table_start_draw_flag + 6 * i)
@@ -915,7 +915,7 @@ def generate_excel(report,
             line.add_data(line_data, titles_from_data=True)
             line.set_categories(labels)
             line_data = line.series[0]
-            line_data.marker.symbol = "circle"
+            line_data.marker.symbol = "auto"
             line_data.smooth = True
             line.x_axis.crosses = 'min'
             line.height = 8.25