+++ myems-api/excelexporters/tenantload.py
@@ -491,14 +491,14 @@ def generate_excel(report,
                     line.add_data(datas, titles_from_data=True)
                     line.set_categories(labels)
                     line_data = line.series[0]
-                    line_data.marker.symbol = "circle"
+                    line_data.marker.symbol = "auto"
                     line_data.smooth = True
                     line.x_axis.crosses = 'min'
                     line.height = 8.25
                     line.width = 24
                     line.dLbls = DataLabelList()
                     line.dLbls.dLblPos = 't'
-                    line.dLbls.showVal = True
+                    line.dLbls.showVal = False
                     ws.add_chart(line, "B" + str(current_chart_row_number))
                     current_chart_row_number += 6
                     current_chart_col_number += 1
@@ -513,14 +513,14 @@ def generate_excel(report,
                     line.add_data(datas, titles_from_data=True)
                     line.set_categories(labels)
                     line_data = line.series[0]
-                    line_data.marker.symbol = "circle"
+                    line_data.marker.symbol = "auto"
                     line_data.smooth = True
                     line.x_axis.crosses = 'min'
                     line.height = 8.25
                     line.width = 24
                     line.dLbls = DataLabelList()
                     line.dLbls.dLblPos = 't'
-                    line.dLbls.showVal = True
+                    line.dLbls.showVal = False
                     ws.add_chart(line, "B" + str(current_chart_row_number))
                     current_chart_row_number += 6
                     current_chart_col_number += 1
@@ -731,14 +731,14 @@ def generate_excel(report,
                         line.add_data(reporting_line_data, titles_from_data=True)
                         line.set_categories(labels)
                         for j in range(len(line.series)):
-                            line.series[j].marker.symbol = "circle"
+                            line.series[j].marker.symbol = "auto"
                             line.series[j].smooth = True
                         line.x_axis.crosses = 'min'
                         line.height = 8.25
                         line.width = 24
                         line.dLbls = DataLabelList()
                         line.dLbls.dLblPos = 't'
-                        line.dLbls.showVal = True
+                        line.dLbls.showVal = False
                         line.dLbls.showPercent = False
                         chart_col = 'B'
                         chart_cell = chart_col + str(current_chart_row_number)
@@ -772,14 +772,14 @@ def generate_excel(report,
                         line.add_data(reporting_line_data, titles_from_data=True)
                         line.set_categories(labels)
                         for j in range(len(line.series)):
-                            line.series[j].marker.symbol = "circle"
+                            line.series[j].marker.symbol = "auto"
                             line.series[j].smooth = True
                         line.x_axis.crosses = 'min'
                         line.height = 8.25
                         line.width = 24
                         line.dLbls = DataLabelList()
                         line.dLbls.dLblPos = 't'
-                        line.dLbls.showVal = True
+                        line.dLbls.showVal = False
                         line.dLbls.showPercent = False
                         chart_col = 'B'
                         chart_cell = chart_col + str(current_chart_row_number)
@@ -949,7 +949,7 @@ def generate_excel(report,
             line.add_data(line_data, titles_from_data=True)
             line.set_categories(labels)
             line_data = line.series[0]
-            line_data.marker.symbol = "circle"
+            line_data.marker.symbol = "auto"
             line_data.smooth = True
             line.x_axis.crosses = 'min'
             line.height = 8.25