+++ myems-api/excelexporters/virtualmeterenergy.py
@@ -355,14 +355,14 @@ def generate_excel(report,
                 line.add_data(line_data, titles_from_data=True)
                 line.set_categories(labels)
                 line_data = line.series[0]
-                line_data.marker.symbol = "circle"
+                line_data.marker.symbol = "auto"
                 line_data.smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(table_start_draw_flag)
@@ -520,14 +520,14 @@ def generate_excel(report,
                 line.add_data(reporting_line_data, titles_from_data=True)
                 line.set_categories(labels)
                 for j in range(len(line.series)):
-                    line.series[j].marker.symbol = "circle"
+                    line.series[j].marker.symbol = "auto"
                     line.series[j].smooth = True
                 line.x_axis.crosses = 'min'
                 line.height = 8.25
                 line.width = 24
                 line.dLbls = DataLabelList()
                 line.dLbls.dLblPos = 't'
-                line.dLbls.showVal = True
+                line.dLbls.showVal = False
                 line.dLbls.showPercent = False
                 chart_col = 'B'
                 chart_cell = chart_col + str(table_start_draw_flag)
@@ -698,7 +698,7 @@ def generate_excel(report,
             line.add_data(line_data, titles_from_data=True)
             line.set_categories(labels)
             line_data = line.series[0]
-            line_data.marker.symbol = "circle"
+            line_data.marker.symbol = "auto"
             line_data.smooth = True
             line.x_axis.crosses = 'min'
             line.height = 8.25